refactored classes into battleship module
